requirementsLab.txt
cffi==1.14.0
cryptography==36.0.1
imageio==2.9.0
joblib==0.17.0
opencv-python-headless
pandas==1.3.3
Pillow==7.2.0
matplotlib==3.5
pydantic==1.10.2
PyJWT==2.6.0
PyYAML==5.3.1
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.6.2
seaborn==0.11.0
tqdm
numpy==1.21.6
QtPy==2.3.0
Pyside2